infra/lambda_src/jogo/requirements.txt
flask
flask_cors
amazon-braket-sdk